Adds a -recognize method to UIGestureRecognizer
This allows the user to trigger gesture recognizers programmatically
under test.

Adds -swipe, -tap and -pinch helper methods to UIView to allow for
the triggering of the appropriate UIGestureRecognizers (if they have
been added to the view).

Adds a whitelisting method to the UIGestureRecognizer extension
category, allowing us to only test the targets that we want to test,
and not hold references to any internal objects in UIKit.

We've not implemented the full set of standard UIGestureRecognizers,
but this should be fairly straightforward given the code we have added.